Checking project update permissions.

+++ impactPathways/src/main/java/org/cgiar/ccafs/ap/action/projects/ProjectCaseStudiesAction.java
@@ -201,54 +201,57 @@ public class ProjectCaseStudiesAction extends BaseAction {
 
   @Override
   public String save() {
-    caseStudiesPreview = caseStudieManager.getCaseStudysByProject(projectID);
-    List<CaseStudieIndicators> indicators;
-    CaseStudieIndicators caseStudieIndicator;
-    if (caseStudiesPreview != null) {
-      for (CasesStudies caseStudie : caseStudiesPreview) {
-        if (!project.getCaseStudies().contains(caseStudie)) {
-          caseStudieManager.deleteCaseStudy(caseStudie.getId(), this.getCurrentUser(), this.getJustification());
-        }
-      }
-    }
-
-    for (CasesStudies caseStudie : project.getCaseStudies()) {
-
-      if (caseStudie.getMyFile() != null) {
-
-        FileManager.deleteFile(this.getCaseStudyPath() + caseStudie.getFile());
-        FileManager.copyFile(caseStudie.getMyFile(), this.getCaseStudyPath() + caseStudie.getMyFileFileName());
-        caseStudie.setFile(caseStudie.getMyFileFileName());
-      }
-      indicators = new ArrayList<>();
-
-
-      if (caseStudie.getCaseStudyIndicatorsIds() != null) {
-        for (String indicator : caseStudie.getCaseStudyIndicatorsIds()) {
-          caseStudieIndicator = new CaseStudieIndicators();
-          caseStudieIndicator.setCasesStudies(caseStudie);
-          caseStudieIndicator.setIdIndicator(Integer.parseInt(indicator));
-          indicators.add(caseStudieIndicator);
+    if (this.hasProjectPermission("update", project.getId())) {
+      caseStudiesPreview = caseStudieManager.getCaseStudysByProject(projectID);
+      List<CaseStudieIndicators> indicators;
+      CaseStudieIndicators caseStudieIndicator;
+      if (caseStudiesPreview != null) {
+        for (CasesStudies caseStudie : caseStudiesPreview) {
+          if (!project.getCaseStudies().contains(caseStudie)) {
+            caseStudieManager.deleteCaseStudy(caseStudie.getId(), this.getCurrentUser(), this.getJustification());
+          }
         }
       }
 
-      caseStudie.setIsActive(true);
-      caseStudie.setCaseStudieIndicatorses(new HashSet<>(indicators));
-      if (caseStudie.getId() != null) {
-        caseStudieManager.saveCaseStudy(projectID, caseStudie, this.getCurrentUser(), this.getJustification());
-      }
+      for (CasesStudies caseStudie : project.getCaseStudies()) {
 
+        if (caseStudie.getMyFile() != null) {
+
+          FileManager.deleteFile(this.getCaseStudyPath() + caseStudie.getFile());
+          FileManager.copyFile(caseStudie.getMyFile(), this.getCaseStudyPath() + caseStudie.getMyFileFileName());
+          caseStudie.setFile(caseStudie.getMyFileFileName());
+        }
+        indicators = new ArrayList<>();
+
+
+        if (caseStudie.getCaseStudyIndicatorsIds() != null) {
+          for (String indicator : caseStudie.getCaseStudyIndicatorsIds()) {
+            caseStudieIndicator = new CaseStudieIndicators();
+            caseStudieIndicator.setCasesStudies(caseStudie);
+            caseStudieIndicator.setIdIndicator(Integer.parseInt(indicator));
+            indicators.add(caseStudieIndicator);
+          }
+        }
+
+        caseStudie.setIsActive(true);
+        caseStudie.setCaseStudieIndicatorses(new HashSet<>(indicators));
+        if (caseStudie.getId() != null) {
+          caseStudieManager.saveCaseStudy(projectID, caseStudie, this.getCurrentUser(), this.getJustification());
+        }
+
+      }
+      // Get the validation messages and append them to the save message
+      Collection<String> messages = this.getActionMessages();
+      if (!messages.isEmpty()) {
+        String validationMessage = messages.iterator().next();
+        this.setActionMessages(null);
+        this.addActionWarning(this.getText("saving.saved") + validationMessage);
+      } else {
+        this.addActionMessage(this.getText("saving.saved"));
+      }
+      return SUCCESS;
     }
-    // Get the validation messages and append them to the save message
-    Collection<String> messages = this.getActionMessages();
-    if (!messages.isEmpty()) {
-      String validationMessage = messages.iterator().next();
-      this.setActionMessages(null);
-      this.addActionWarning(this.getText("saving.saved") + validationMessage);
-    } else {
-      this.addActionMessage(this.getText("saving.saved"));
-    }
-    return SUCCESS;
+    return NOT_AUTHORIZED;
   }
